refactor(app): simplify transfer consolidation split
